requirements.txt
akismet==0.2.0
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.4.0
billiard==3.3.0.20
blessings==1.6
blinker==1.4
bpython==0.14.2
celery==3.1.18
celery-with-redis==3.0
cffi==1.2.1
cryptography==1.0.1
cssmin==0.2.0
cssselect==0.9.1
curtsies==0.1.19
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.2
facebook-sdk==1.0.0
Flask==0.10.1
Flask-Assets==0.11
Flask-Cache==0.13.1
Flask-Celery==2.4.3
Flask-DebugToolbar==0.10.0
Flask-ErrorMail==0.2.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
flask-mongoengine==0.7.1
Flask-OAuth==0.12
Flask-OpenID==1.2.4
Flask-Principal==0.4.0
Flask-PyMongo==0.3.1
Flask-Rauth==0.3.2
Flask-Script==2.0.5
# Flask-Security==1.7.4
-e git+https://github.com/guoqiao/flask-security#egg=Flask-Security
Flask-Social==1.6.2
Flask-SQLAlchemy==2.0
Flask-Uploads==0.1.3
Flask-WTF==0.12
gdata==2.0.18
google-api-python-client==1.4.2
greenlet==0.4.9
httplib2==0.9.1
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.26
linkedin==0.1.5
Markdown==2.6.2
MarkupSafe==0.23
meld3==1.0.2
mongoengine==0.10.0
oauth==1.0.1
oauth2==1.9.0.post1
oauth2client==1.5.0
oauthlib==1.0.3
openid-mongodb==0.1
paramiko==1.15.2
passlib==1.6.5
Pillow==2.9.0
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
pycrypto==2.6.1
Pygments==2.0.2
pymongo==3.0.3
pyOpenSSL==0.15.1
python-dateutil==2.4.2
python-gflags==2.0
python-openid==2.2.5
python-slugify==1.1.3
python-twitter==2.2
pytz==2015.4
rauth==0.4.17
rawes==0.5.5
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
rsa==3.2
sh==1.11
six==1.9.0
SQLAlchemy==1.0.8
ssh==1.8.0
supervisor==3.1.3
thrift==0.9.2
Unidecode==0.4.18
uritemplate==0.6
virtualenv==13.1.2
webassets==0.11.1
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
yos-social-sdk==0.0.2
jsmin==2.1.6
